Pagination & React & MultiLanguage/main/api/serializers.py
from operator import mod
from rest_framework import serializers
from main import models


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Product
        fields = '__all__'
