Pagination & React & MultiLanguage/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.3.0
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
toposort==1.7
